move headers from include/DA to include/dace move CMake option handling to config.h file to make options persistent (i.e. #ifdefs in headers also reflect configuration options in deployed installations)
