+++ core/dacemath.c
@@ -37,8 +37,9 @@
 #include <math.h>
 #include <stdlib.h>
 
-#include "DA/dacebase.h"
-#include "DA/daceaux.h"
+#include "dace/config.h"
+#include "dace/dacebase.h"
+#include "dace/daceaux.h"
 
 // define various math constants in case they have not been defined by math.h
 // these are non-standard C, but most C libraries have them
@@ -109,7 +110,7 @@ void daceMultiply(const DACEDA *ina, const DACEDA *inb, DACEDA *inc)
 {
 // These should use thread local storage (TLS) for multithread safe implementations
 // see https://en.wikipedia.org/wiki/Thread-local_storage
-#ifdef DACE_STATIC_MEMORY
+#if DACE_MEMORY_MODEL == DACE_MEMORY_STATIC
     static DACE_THREAD_LOCAL double cc[DACE_STATIC_NMMAX] = {0};
     static DACE_THREAD_LOCAL extended_monomial emb[DACE_STATIC_NMMAX];
     static DACE_THREAD_LOCAL extended_monomial *ipbeg[DACE_STATIC_NOMAX+1];
@@ -712,7 +713,7 @@ void dacePowerDouble(const DACEDA *ina, const double p, DACEDA *inc)
         return;
     }
 
-#ifdef DACE_STATIC_MEMORY
+#if DACE_MEMORY_MODEL == DACE_MEMORY_STATIC
     double xf[DACE_STATIC_NOMAX+1];
 #else
     double *xf = (double*) dacecalloc(DACECom_t.nocut+1, sizeof(double));
@@ -724,7 +725,7 @@ void dacePowerDouble(const DACEDA *ina, const double p, DACEDA *inc)
 
     daceDivideDouble(ina, a0, inc);     // more accurate than including a0 in series (uses non-linear part in EvaluateSeries)
     daceEvaluateSeries(inc, xf, inc);
-#ifndef DACE_STATIC_MEMORY
+#if DACE_MEMORY_MODEL != DACE_MEMORY_STATIC
    dacefree(xf);
 #endif
 }
@@ -828,7 +829,7 @@ void daceRoot(const DACEDA *ina, const int np, DACEDA *inc)
     }
 
     double cr = 1.0/np;
-#ifdef DACE_STATIC_MEMORY
+#if DACE_MEMORY_MODEL == DACE_MEMORY_STATIC
     double xf[DACE_STATIC_NOMAX+1];
 #else
     double *xf = (double*) dacecalloc(DACECom_t.nocut+1, sizeof(double));
@@ -843,7 +844,7 @@ void daceRoot(const DACEDA *ina, const int np, DACEDA *inc)
 
     daceDivideDouble(ina, a0, inc);     // more accurate than including a0 in series (uses non-linear part in EvaluateSeries)
     daceEvaluateSeries(inc, xf, inc);
-#ifndef DACE_STATIC_MEMORY
+#if DACE_MEMORY_MODEL != DACE_MEMORY_STATIC
    dacefree(xf);
 #endif
 }
@@ -898,7 +899,7 @@ void daceMultiplicativeInverse0(const DACEDA *ina, DACEDA *inc, const double a0)
 {
     daceDivideDouble(ina, a0, inc);
 
-#ifdef DACE_STATIC_MEMORY
+#if DACE_MEMORY_MODEL == DACE_MEMORY_STATIC
     double xf[DACE_STATIC_NOMAX+1];
 #else
     double *xf = (double*) dacecalloc(DACECom_t.nocut+1, sizeof(double));
@@ -909,7 +910,7 @@ void daceMultiplicativeInverse0(const DACEDA *ina, DACEDA *inc, const double a0)
         xf[i] = -xf[i-1];
 
     daceEvaluateSeries(inc, xf, inc);
-#ifndef DACE_STATIC_MEMORY
+#if DACE_MEMORY_MODEL != DACE_MEMORY_STATIC
     dacefree(xf);
 #endif
 }
@@ -981,7 +982,7 @@ void daceHypotenuse(const DACEDA *ina, const DACEDA *inb, DACEDA *inc)
  */
 void daceExponential(const DACEDA *ina, DACEDA *inc)
 {
-#ifdef DACE_STATIC_MEMORY
+#if DACE_MEMORY_MODEL == DACE_MEMORY_STATIC
     double xf[DACE_STATIC_NOMAX+1];
 #else
     double* xf = (double*) dacecalloc(DACECom_t.nocut+1, sizeof(double));
@@ -992,7 +993,7 @@ void daceExponential(const DACEDA *ina, DACEDA *inc)
         xf[i] = xf[i-1]/i;
 
     daceEvaluateSeries(ina, xf, inc);
-#ifndef DACE_STATIC_MEMORY
+#if DACE_MEMORY_MODEL != DACE_MEMORY_STATIC
     dacefree(xf);
 #endif
 }
@@ -1012,7 +1013,7 @@ void daceLogarithm(const DACEDA *ina, DACEDA *inc)
         return;
     }
 
-#ifdef DACE_STATIC_MEMORY
+#if DACE_MEMORY_MODEL == DACE_MEMORY_STATIC
     double xf[DACE_STATIC_NOMAX+1];
 #else
     double* xf = (double*) dacecalloc(DACECom_t.nocut+1, sizeof(double));
@@ -1027,7 +1028,7 @@ void daceLogarithm(const DACEDA *ina, DACEDA *inc)
     }
 
     daceEvaluateSeries(inc, xf, inc);
-#ifndef DACE_STATIC_MEMORY
+#if DACE_MEMORY_MODEL != DACE_MEMORY_STATIC
     dacefree(xf);
 #endif
 }
@@ -1078,7 +1079,7 @@ void daceLogarithm2(const DACEDA *ina, DACEDA *inc)
  */
 void daceSine(const DACEDA *ina, DACEDA *inc)
 {
-#ifdef DACE_STATIC_MEMORY
+#if DACE_MEMORY_MODEL == DACE_MEMORY_STATIC
     double xf[DACE_STATIC_NOMAX+1];
 #else
     double* xf = (double*) dacecalloc(DACECom_t.nocut+1, sizeof(double));
@@ -1093,7 +1094,7 @@ void daceSine(const DACEDA *ina, DACEDA *inc)
     }
 
     daceEvaluateSeries(ina, xf, inc);
-#ifndef DACE_STATIC_MEMORY
+#if DACE_MEMORY_MODEL != DACE_MEMORY_STATIC
     dacefree(xf);
 #endif
 }
@@ -1105,7 +1106,7 @@ void daceSine(const DACEDA *ina, DACEDA *inc)
  */
 void daceCosine(const DACEDA *ina, DACEDA *inc)
 {
-#ifdef DACE_STATIC_MEMORY
+#if DACE_MEMORY_MODEL == DACE_MEMORY_STATIC
     double xf[DACE_STATIC_NOMAX+1];
 #else
     double* xf = (double*) dacecalloc(DACECom_t.nocut+1, sizeof(double));
@@ -1120,7 +1121,7 @@ void daceCosine(const DACEDA *ina, DACEDA *inc)
     }
 
     daceEvaluateSeries(ina, xf, inc);
-#ifndef DACE_STATIC_MEMORY
+#if DACE_MEMORY_MODEL != DACE_MEMORY_STATIC
     dacefree(xf);
 #endif
 }
@@ -1198,7 +1199,7 @@ void daceArcCosine(const DACEDA *ina, DACEDA *inc)
 void daceArcTangent(const DACEDA *ina, DACEDA *inc)
 {
     DACEDA iarg;
-#ifdef DACE_STATIC_MEMORY
+#if DACE_MEMORY_MODEL == DACE_MEMORY_STATIC
     double xf[DACE_STATIC_NOMAX+1] = {0};
 #else
     double* xf = (double*) dacecalloc(DACECom_t.nocut+1, sizeof(double));
@@ -1221,7 +1222,7 @@ void daceArcTangent(const DACEDA *ina, DACEDA *inc)
 
     daceEvaluateSeries(&iarg, xf, inc);
     daceFreeDA(&iarg);
-#ifndef DACE_STATIC_MEMORY
+#if DACE_MEMORY_MODEL != DACE_MEMORY_STATIC
     dacefree(xf);
 #endif
 }
@@ -1283,7 +1284,7 @@ void daceArcTangent2(const DACEDA *ina, const DACEDA *inb, DACEDA *inc)
  */
 void daceHyperbolicSine(const DACEDA *ina, DACEDA *inc)
 {
-#ifdef DACE_STATIC_MEMORY
+#if DACE_MEMORY_MODEL == DACE_MEMORY_STATIC
     double xf[DACE_STATIC_NOMAX+1];
 #else
     double* xf = (double*) dacecalloc(DACECom_t.nocut+1, sizeof(double));
@@ -1299,7 +1300,7 @@ void daceHyperbolicSine(const DACEDA *ina, DACEDA *inc)
     }
 
     daceEvaluateSeries(ina, xf, inc);
-#ifndef DACE_STATIC_MEMORY
+#if DACE_MEMORY_MODEL != DACE_MEMORY_STATIC
     dacefree(xf);
 #endif
 }
@@ -1311,7 +1312,7 @@ void daceHyperbolicSine(const DACEDA *ina, DACEDA *inc)
  */
 void daceHyperbolicCosine(const DACEDA *ina, DACEDA *inc)
 {
-#ifdef DACE_STATIC_MEMORY
+#if DACE_MEMORY_MODEL == DACE_MEMORY_STATIC
     double xf[DACE_STATIC_NOMAX+1];
 #else
     double* xf = (double*) dacecalloc(DACECom_t.nocut+1, sizeof(double));
@@ -1327,7 +1328,7 @@ void daceHyperbolicCosine(const DACEDA *ina, DACEDA *inc)
     }
 
     daceEvaluateSeries(ina, xf, inc);
-#ifndef DACE_STATIC_MEMORY
+#if DACE_MEMORY_MODEL != DACE_MEMORY_STATIC
     dacefree(xf);
 #endif
 }
@@ -1423,7 +1424,7 @@ void daceHyperbolicArcTangent(const DACEDA *ina, DACEDA *inc)
  */
 void daceErrorFunction(const DACEDA *ina, DACEDA *inc)
 {
-#ifdef DACE_STATIC_MEMORY
+#if DACE_MEMORY_MODEL == DACE_MEMORY_STATIC
     double xf[DACE_STATIC_NOMAX+1];
 #else
     double* xf = (double*) dacecalloc(DACECom_t.nocut+1, sizeof(double));
@@ -1446,7 +1447,7 @@ void daceErrorFunction(const DACEDA *ina, DACEDA *inc)
     }
 
     daceEvaluateSeries(ina, xf, inc);
-#ifndef DACE_STATIC_MEMORY
+#if DACE_MEMORY_MODEL != DACE_MEMORY_STATIC
     dacefree(xf);
 #endif
 }
@@ -1458,7 +1459,7 @@ void daceErrorFunction(const DACEDA *ina, DACEDA *inc)
  */
 void daceComplementaryErrorFunction(const DACEDA *ina, DACEDA *inc)
 {
-#ifdef DACE_STATIC_MEMORY
+#if DACE_MEMORY_MODEL == DACE_MEMORY_STATIC
     double xf[DACE_STATIC_NOMAX+1];
 #else
     double* xf = (double*) dacecalloc(DACECom_t.nocut+1, sizeof(double));
@@ -1481,7 +1482,7 @@ void daceComplementaryErrorFunction(const DACEDA *ina, DACEDA *inc)
     }
 
     daceEvaluateSeries(ina, xf, inc);
-#ifndef DACE_STATIC_MEMORY
+#if DACE_MEMORY_MODEL != DACE_MEMORY_STATIC
     dacefree(xf);
 #endif
 }
@@ -1493,7 +1494,7 @@ void daceComplementaryErrorFunction(const DACEDA *ina, DACEDA *inc)
  */
 void daceBesselJFunction(const DACEDA *ina, const int n, DACEDA *inc)
 {
-#ifdef DACE_STATIC_MEMORY
+#if DACE_MEMORY_MODEL == DACE_MEMORY_STATIC
     double bz[2*DACE_STATIC_NOMAX+1];
 #else
     double* bz = (double*) dacecalloc(2*DACECom_t.nocut+1, sizeof(double));
@@ -1504,7 +1505,7 @@ void daceBesselJFunction(const DACEDA *ina, const int n, DACEDA *inc)
         bz[i] = jn(n-DACECom_t.nocut+i, a0);        // MSVC recommends using _jn, but not portable
 
     daceEvaluateBesselFunction(ina, bz, inc);
-#ifndef DACE_STATIC_MEMORY
+#if DACE_MEMORY_MODEL != DACE_MEMORY_STATIC
     dacefree(bz);
 #endif
 }
@@ -1524,7 +1525,7 @@ void daceBesselYFunction(const DACEDA *ina, const int n, DACEDA *inc)
         return;
     }
 
-#ifdef DACE_STATIC_MEMORY
+#if DACE_MEMORY_MODEL == DACE_MEMORY_STATIC
     double bz[2*DACE_STATIC_NOMAX+1];
 #else
     double* bz = (double*) dacecalloc(2*DACECom_t.nocut+1, sizeof(double));
@@ -1534,7 +1535,7 @@ void daceBesselYFunction(const DACEDA *ina, const int n, DACEDA *inc)
         bz[i] = yn(n-DACECom_t.nocut+i, a0);		// MSVC recommends using _yn, but not portable
 
     daceEvaluateBesselFunction(ina, bz, inc);
-#ifndef DACE_STATIC_MEMORY
+#if DACE_MEMORY_MODEL != DACE_MEMORY_STATIC
     dacefree(bz);
 #endif
 }
@@ -1547,7 +1548,7 @@ void daceBesselYFunction(const DACEDA *ina, const int n, DACEDA *inc)
  */
 void daceEvaluateBesselFunction(const DACEDA *ina, const double bz[], DACEDA *inc)
 {
-#ifdef DACE_STATIC_MEMORY
+#if DACE_MEMORY_MODEL == DACE_MEMORY_STATIC
     double xf[DACE_STATIC_NOMAX+1];
     double binomial[DACE_STATIC_NOMAX+1];
 #else
@@ -1578,7 +1579,7 @@ void daceEvaluateBesselFunction(const DACEDA *ina, const double bz[], DACEDA *in
     }
 
     daceEvaluateSeries(ina, xf, inc);
-#ifndef DACE_STATIC_MEMORY
+#if DACE_MEMORY_MODEL != DACE_MEMORY_STATIC
     dacefree(binomial);
     dacefree(xf);
 #endif
