+++ interfaces/cxx/include/dace/DAFormatter.h
@@ -33,7 +33,7 @@
 #include <vector>
 #include <string>
 
-#include "DA/Def.h"
+#include "dace/Def.h"
 
 namespace DACE{
