+++ interfaces/cxx/include/dace/Monomial.h
@@ -34,7 +34,7 @@
 #include <string>
 #include <ostream>
 
-#include "DA/Def.h"
+#include "dace/Def.h"
 
 namespace DACE{
