+++ interfaces/cxx/include/dace/compiledDA.h
@@ -32,7 +32,7 @@
 // C++ stdlib classes used in this public interface
 #include <vector>
 
-#include "DA/Def.h"
+#include "dace/Def.h"
 
 namespace DACE{
